Support clean restarts
We currently run a shell as PID 1, and then use sudo to run Postgres.
Running a shell is bad because it means all signals are lost (including
the SIGTERM sent by `docker restart`), and using sudo is bad because we
can't really ever know for sure what happens to signals or the
environment.

All in all, this patch ensures PG runs as PID 1 and adds a test to make
sure it restarts cleanly on a `docker restart`.

This also increases the Postgres log level to INFO, so that we can
capture log messages to know what Postgres is doing (and also because
it's good to hear about warnings and so on).